pyproject.toml
[project]
name = "nl2sql-agent-adk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.14.0",
    "google-cloud-aiplatform>=1.70.0",
    "python-dotenv>=1.0.0",
]
